Move route specific specs to own directory
